+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Update file in your sub-theme:
 - Add paragraphs base module
 - Switch to classy base theme
 - Add block configuration for: slim menu and small prints menu
+- Add paragraph overlay
 
 # Summary 8.x.0.11
 ## Release note